Refactor code
The original vocal.py is being renamed to session_manager.py, and the code is being split into audio_source_handlers.py, audio_controls.py, queue_view.py and server_session.py. All imports are being modified too.
