+++ bot/audio_controls.py
@@ -1,0 +1,65 @@
+from bot.server_session import ServerSession
+from bot.utils import update_active_servers
+from bot.session_manager import server_sessions
+
+
+async def seek_playback(guild_id: str, position: int):
+    guild_id = int(guild_id)
+    if guild_id not in server_sessions:
+        return False
+
+    session: ServerSession = server_sessions[guild_id]
+    return await session.seek(position)
+
+
+async def toggle_loop(guild_id: str, mode: str):
+    guild_id = int(guild_id)
+    if guild_id not in server_sessions:
+        return False
+
+    session: ServerSession = server_sessions[guild_id]
+
+    return await session.toggle_loop(mode)
+
+
+async def skip_track(guild_id: str):
+    guild_id = int(guild_id)
+    if guild_id not in server_sessions:
+        return False
+
+    session: ServerSession = server_sessions[guild_id]
+    return await session.skip_track(session.last_context)
+
+
+async def previous_track(guild_id: str):
+    guild_id = int(guild_id)
+    if guild_id not in server_sessions:
+        return False
+
+    session: ServerSession = server_sessions[guild_id]
+    return await session.previous_track(session.last_context)
+
+
+async def shuffle_queue(guild_id: str, is_active: bool):
+    guild_id = int(guild_id)
+    if guild_id not in server_sessions:
+        return False
+
+    session: ServerSession = server_sessions[guild_id]
+    success = await session.shuffle_queue(is_active)
+
+    # Update the queue for all connected clients
+    await update_active_servers(session.bot, server_sessions)
+
+    return success
+
+
+async def set_volume(guild_id: str, volume: int):
+    guild_id = int(guild_id)
+    if guild_id not in server_sessions:
+        return False
+
+    session: ServerSession = server_sessions[guild_id]
+    session.voice_client.source.volume = volume / 100
+    session.volume = volume
+    return True
